move XcodeSDKPath in XcodeSDK.h
