+++ lldb/include/lldb/Utility/XcodeSDK.h
@@ -9,6 +9,7 @@
 #ifndef LLDB_UTILITY_XCODESDK_H
 #define LLDB_UTILITY_XCODESDK_H
 
+#include "lldb/Utility/FileSpec.h"
 #include "lldb/lldb-forward.h"
 #include "llvm/ADT/StringRef.h"
 #include "llvm/Support/VersionTuple.h"
@@ -93,6 +94,31 @@ public:
   static XcodeSDK::Type GetSDKTypeForTriple(const llvm::Triple &triple);
 
   static std::string FindXcodeContentsDirectoryInPath(llvm::StringRef path);
+};
+
+/// An abstraction which groups an XcodeSDK with its parsed path.
+class XcodeSDKPath {
+  XcodeSDK m_sdk;
+  FileSpec m_sysroot;
+
+public:
+  /// Default constructor, constructs an empty sdk with an empty path.
+  XcodeSDKPath() = default;
+  XcodeSDKPath(XcodeSDK sdk, FileSpec sysroot)
+      : m_sdk(std::move(sdk)), m_sysroot(std::move(sysroot)) {}
+  XcodeSDKPath(std::string name, FileSpec sysroot)
+      : m_sdk(XcodeSDK(std::move(name))), m_sysroot(std::move(sysroot)) {}
+
+  bool operator==(const XcodeSDKPath &other) const;
+  bool operator!=(const XcodeSDKPath &other) const;
+
+  XcodeSDK TakeSDK() const;
+  const FileSpec &GetSysroot() const { return m_sysroot; }
+  llvm::StringRef GetString() const { return m_sdk.GetString(); }
+  XcodeSDK::Type GetType() const { return m_sdk.GetType(); }
+
+  void Merge(const XcodeSDKPath &other);
+  bool IsAppleInternalSDK() const { return m_sdk.IsAppleInternalSDK(); }
 };
 
 } // namespace lldb_private
